+++ samcli/commands/check/resources/api_gateway.py
@@ -1,7 +1,7 @@
-from .TemplateResource import TemplateResource
+from samcli.commands.check.resources.template_resource import TemplateResource
 
 
-class DynamoDB(TemplateResource):
+class ApiGateway(TemplateResource):
     def __init__(self, resource_object, resource_type, resource_name):
         super().__init__(resource_object, resource_type, resource_name)
         self.tps = None
