+++ samcli/commands/check/resources/event_source_mapping.py
@@ -1,6 +1,6 @@
-from .TemplateResource import TemplateResource
+from samcli.commands.check.resources.template_resource import TemplateResource
 
 
-class LambdaFunctionPermission(TemplateResource):
+class EventSourceMapping(TemplateResource):
     def __init__(self, resource_object, resource_type, resource_name):
         super().__init__(resource_object, resource_type, resource_name)
