+++ samcli/commands/check/resources/i_am_role.py
@@ -1,6 +1,7 @@
-from .TemplateResource import TemplateResource
+from samcli.commands.check.resources.template_resource import TemplateResource
 
 
-class EventSourceMapping(TemplateResource):
+class IAMRole(TemplateResource):
     def __init__(self, resource_object, resource_type, resource_name):
         super().__init__(resource_object, resource_type, resource_name)
+        self.policies = []
