+++ samcli/commands/check/resources/lambda_function_permission.py
@@ -1,7 +1,6 @@
-from .TemplateResource import TemplateResource
+from samcli.commands.check.resources.template_resource import TemplateResource
 
 
-class IAMRole(TemplateResource):
+class LambdaFunctionPermission(TemplateResource):
     def __init__(self, resource_object, resource_type, resource_name):
         super().__init__(resource_object, resource_type, resource_name)
-        self.policies = []
