+++ samcli/commands/check/resources/template_resource.py
