+++ samcli/commands/check/resources/warning.py
